+++ msgcheck/msgcheck.py
@@ -66,8 +66,9 @@ The script returns:
                         help='do not check compilation of file')
     parser.add_argument('-f', '--fuzzy', action='store_true',
                         help='check fuzzy strings')
-    parser.add_argument('-n', '--skip-noqa', action='store_true',
-                        help='do not check "noqa"-commented lines')
+    parser.add_argument('-n', '--check-noqa', action='store_true',
+                        help='check "noqa"-commented lines (they are skipped '
+                        'by default)')
     parser.add_argument('-l', '--no-lines', action='store_true',
                         help='do not check number of lines')
     parser.add_argument('-p', '--no-punct', action='store_true',
@@ -118,8 +119,9 @@ def msgcheck_check_files(args):
     """Check files."""
     # create checker and set boolean options
     po_check = PoCheck()
-    for option in ('no_compile', 'fuzzy', 'skip_noqa', 'no_lines', 'no_punct',
-                   'no_whitespace', 'no_whitespace_eol', 'extract'):
+    for option in ('no_compile', 'fuzzy', 'check_noqa', 'no_lines',
+                   'no_punct', 'no_whitespace', 'no_whitespace_eol',
+                   'extract'):
         if args.__dict__[option]:
             po_check.set_check(option.lstrip('no_'),
                                not option.startswith('no_'))